first go at bringing in wordpress with graph ql and gridsome 0.1
