--- dima-portfolio/gridsome.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  //plugins: [
-  // {
-  //   use: '@gridsome/source-wordpress',
-  //   options: {
-  //     baseUrl: 'YOUR_WORDPRESS_URL'
-  //   }
-  // }
-  //]
-};
