+++ dimapress/README.md
@@ -1,0 +1,35 @@
+# WordPress starter for Gridsome
+
+## Install
+`gridsome create my-gridsome-project wordpress`
+
+## Guide
+
+Add your WordPress URL to the plugin options.
+
+```js
+// gridsome.config.js
+
+module.exports = {
+  plugins: [
+    {
+      use: '@gridsome/source-wordpress',
+      options: {
+        baseUrl: 'YOUR_WEBSITE_URL', // required
+        typeName: 'WordPress', // GraphQL schema name (Optional)
+        perPage: 100, // How many posts to load from server per request (Optional)
+        concurrent: 10, // How many requests to run simultaneously (Optional)
+        routes: {
+          post: '/:year/:month/:day/:slug', //adds route for "post" post type (Optional)
+          post_tag: '/tag/:slug' // adds route for "post_tag" post type (Optional)
+        }
+      }
+    }
+  ]
+}
+
+```
+
+## Included templates
+
+This starter includes basic templates for categories, tags and posts.
