+++ dimapress/src/layouts/Base/Head.vue
