+++ dimapress/src/main.js
@@ -1,0 +1,5 @@
+import DefaultLayout from '~/layouts/Default.vue'
+
+export default function (Vue) {
+  Vue.component('Layout', DefaultLayout)
+}
