+++ dimapress/src/templates/WordPressCategory.vue
@@ -4,14 +4,14 @@
   </Layout>
 </template>
 
-<!--<page-query>
+<page-query>
 query Category ($path: String!) {
   wordPressCategory (path: $path) {
     title
   }
 }
-</page-query> -->
+</page-query>
 
 <script>
-export default {};
+export default {}
 </script>
